+++ Tests/HackleTests/Mock/MockHackleApp.swift
@@ -176,9 +176,9 @@ class MockHackleApp : Mock, HackleAppProtocol {
         return call(remoteConfigWithUserRef, args: (user))
     }
     
-    lazy var setCurrentScreenRef = MockFunction(self, setCurrentScreen as (String, String) -> Void)
-    func setCurrentScreen(screenName: String, className: String) {
-        call(setCurrentScreenRef, args: (screenName, className))
+    lazy var setCurrentScreenRef = MockFunction(self, setCurrentScreen as (Screen) -> Void)
+    func setCurrentScreen(screen: Screen) {
+        call(setCurrentScreenRef, args: (screen))
     }
     
     func allVariationDetails() -> [Int : Decision] {
